Move synthetic_stack to ImageStack
