+++ starfish/test/image/test_imagestack.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from starfish.constants import Indices
-from starfish.test.dataset_fixtures import synthetic_stack
+from starfish.image import ImageStack
 
 
 def test_get_slice_simple_index():
@@ -9,7 +9,7 @@ def test_get_slice_simple_index():
     Retrieve a slice across one of the indices at the end.  For instance, if the dimensions are
     (P, Q0,..., Qn-1, R), slice across either P or R.
     """
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     hyb = 1
     imageslice, axes = stack.get_slice(
         {Indices.HYB: hyb}
@@ -31,7 +31,7 @@ def test_get_slice_middle_index():
     Retrieve a slice across one of the indices in the middle.  For instance, if the dimensions are
     (P, Q0,..., Qn-1, R), slice across one of the Q axes.
     """
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     ch = 1
     imageslice, axes = stack.get_slice(
         {Indices.CH: ch}
@@ -52,7 +52,7 @@ def test_get_slice_range():
     """
     Retrieve a slice across a range of one of the dimensions.
     """
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     zrange = slice(1, 3)
     imageslice, axes = stack.get_slice(
         {Indices.Z: zrange}
@@ -75,7 +75,7 @@ def test_set_slice_simple_index():
     Sets a slice across one of the indices at the end.  For instance, if the dimensions are
     (P, Q0,..., Qn-1, R), sets a slice across either P or R.
     """
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     hyb = 1
     Y, X = stack.tile_shape
 
@@ -92,7 +92,7 @@ def test_set_slice_middle_index():
     Sets a slice across one of the indices in the middle.  For instance, if the dimensions are
     (P, Q0,..., Qn-1, R), slice across one of the Q axes.
     """
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     ch = 1
     Y, X = stack.tile_shape
 
@@ -108,7 +108,7 @@ def test_set_slice_range():
     """
     Sets a slice across a range of one of the dimensions.
     """
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     zrange = slice(1, 3)
     Y, X = stack.tile_shape
