+++ starfish/test/image/test_imagestack_metadata.py
@@ -3,7 +3,11 @@ from typing import Any
 import pytest
 
 from starfish.constants import Indices
-from starfish.test.dataset_fixtures import DEFAULT_NUM_HYB, DEFAULT_NUM_CH, DEFAULT_NUM_Z, synthetic_stack
+from starfish.image import ImageStack
+
+NUM_HYB = 4
+NUM_CH = 2
+NUM_Z = 12
 
 
 def test_metadata():
@@ -19,11 +23,11 @@ def test_metadata():
             }
         }
 
-    stack = synthetic_stack(
-        tile_extras_provider=tile_extras_provider,
+    stack = ImageStack.synthetic_stack(
+        num_hyb=NUM_HYB, num_ch=NUM_CH, num_z=NUM_Z, tile_extras_provider=tile_extras_provider,
     )
     table = stack.tile_metadata
-    assert len(table) == DEFAULT_NUM_HYB * DEFAULT_NUM_CH * DEFAULT_NUM_Z
+    assert len(table) == NUM_HYB * NUM_CH * NUM_Z
 
 
 def test_missing_extras():
@@ -42,11 +46,11 @@ def test_missing_extras():
         else:
             return None
 
-    stack = synthetic_stack(
-        tile_extras_provider=tile_extras_provider,
+    stack = ImageStack.synthetic_stack(
+        num_hyb=NUM_HYB, num_ch=NUM_CH, num_z=NUM_Z, tile_extras_provider=tile_extras_provider,
     )
     table = stack.tile_metadata
-    assert len(table) == DEFAULT_NUM_HYB * DEFAULT_NUM_CH * DEFAULT_NUM_Z
+    assert len(table) == NUM_HYB * NUM_CH * NUM_Z
 
 
 def test_conflict():
@@ -60,8 +64,8 @@ def test_conflict():
             Indices.Z: z,
         }
 
-    stack = synthetic_stack(
-        tile_extras_provider=tile_extras_provider,
+    stack = ImageStack.synthetic_stack(
+        num_hyb=NUM_HYB, num_ch=NUM_CH, num_z=NUM_Z, tile_extras_provider=tile_extras_provider,
     )
     with pytest.raises(ValueError):
         stack.tile_metadata
