+++ starfish/test/test_apply.py
@@ -2,7 +2,8 @@ from copy import deepcopy
 
 import numpy as np
 
-from starfish.test.dataset_fixtures import synthetic_stack, labeled_synthetic_dataset
+from starfish.test.dataset_fixtures import labeled_synthetic_dataset
+from starfish.image import ImageStack
 
 
 def multiply(array, value):
@@ -11,7 +12,7 @@ def multiply(array, value):
 
 def test_apply():
     """test that apply correctly applies a simple function across 2d tiles of a Stack"""
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     assert (stack.numpy_array == 1).all()
     stack.apply(multiply, value=2)
     assert (stack.numpy_array == 2).all()
@@ -19,7 +20,7 @@ def test_apply():
 
 def test_apply_3d():
     """test that apply correctly applies a simple function across 3d volumes of a Stack"""
-    stack = synthetic_stack()
+    stack = ImageStack.synthetic_stack()
     assert np.all(stack.numpy_array == 1)
     stack.apply(multiply, value=4, is_volume=True)
     assert (stack.numpy_array == 4).all()
